# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/paath": {
      "get": {
        "operationId": "PaathController_findAll",
        "summary": "List published Nitya Paath & Shlokas (public)",
        "parameters": [
          {
            "name": "templeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Paath"
        ]
      }
    },
...
    "/api/v1/paath/{id}": {
      "get": {
        "operationId": "PaathController_findById",
        "summary": "Get single published Paath by ID (public)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Paath"
        ]
      }
    },
    "/api/v1/admin/paath": {
      "get": {
        "operationId": "AdminPaathController_adminFindAll",
        "summary": "List all Paath items including drafts (admin)",
        "parameters": [
          {
            "name": "templeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isPublished",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Paath Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "AdminPaathController_create",
        "summary": "Create new Paath / Shloka content (admin)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaathDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Paath Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/paath/{id}": {
      "put": {
        "operationId": "AdminPaathController_update",
        "summary": "Update Paath content (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaathDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Paath Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminPaathController_delete",
        "summary": "Delete Paath content (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Paath Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/paath/{id}/publish": {
      "put": {
        "operationId": "AdminPaathController_setPublishStatus",
        "summary": "Publish or unpublish Paath content (admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Paath Admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Temple Digital Platform API",
    "description": "Comprehensive RESTful Backend API for Temple Digital Platform (Devotee & Admin Panel)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePaathDto": {
        "type": "object",
        "properties": {
          "templeId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "description": "Title of the Shloka / Paath"
          },
          "sanskritText": {
            "type": "string",
            "description": "Original Sanskrit text / Mantra"
          },
          "transliteration": {
            "type": "string",
            "description": "Roman / Hindi transliteration for chanting"
          },
          "hindiMeaning": {
            "type": "string",
            "description": "Hindi translation and meaning"
          },
          "englishMeaning": {
            "type": "string",
            "description": "English translation and meaning"
          },
          "audioUrl": {
            "type": "string",
            "description": "Audio streaming URL"
          },
          "durationSeconds": {
            "type": "number",
            "description": "Duration in seconds"
          },
          "category": {
            "type": "string",
            "description": "Category e.g. Vedic Mantra, Stotram, Aarti"
          },
          "isPublished": {
            "type": "boolean",
            "default": true
          },
          "displayOrder": {
            "type": "number",
            "default": 0
          }
        },
        "required": [
          "title",
          "sanskritText",
          "transliteration",
          "hindiMeaning"
        ]
      },
      "UpdatePaathDto": {
        "type": "object",
        "properties": {
          "templeId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "description": "Title of the Shloka / Paath"
          },
          "sanskritText": {
            "type": "string",
            "description": "Original Sanskrit text / Mantra"
          },
          "transliteration": {
            "type": "string",
            "description": "Roman / Hindi transliteration for chanting"
          },
          "hindiMeaning": {
            "type": "string",
            "description": "Hindi translation and meaning"
          },
          "englishMeaning": {
            "type": "string",
            "description": "English translation and meaning"
          },
          "audioUrl": {
            "type": "string",
            "description": "Audio streaming URL"
          },
          "durationSeconds": {
            "type": "number",
            "description": "Duration in seconds"
          },
          "category": {
            "type": "string",
            "description": "Category e.g. Vedic Mantra, Stotram, Aarti"
          },
          "isPublished": {
            "type": "boolean",
            "default": true
          },
          "displayOrder": {
            "type": "number",
            "default": 0
          }
        }
      },
...
    }
  }
}